Moved "Functions" static class to its own file Do not use a new Random each time. Random is seeded on creation from the system clock. Two different randoms created close enough together can return identical values. Also it's an unnecessary object aloocation.
